make sizes and offsets int, not uint64 add Value.Interface, to extract an empty interface that can be converted to a regular Go value of the appropriate type, if known.
R=rsc
DELTA=49  (2 added, 0 deleted, 47 changed)
OCL=18526
CL=18526

+++ src/lib/reflect/test.go
@@ -202,7 +202,7 @@ func main() {
 	var name string;
 	var typ reflect.Type;
 	var tag string;
-	var offset uint64;
+	var offset int;
 
 	// Type strings
 	t = reflect.ParseTypeString("", "int8");
